+++ packages/vike-react-apollo/renderer/+config.ts
@@ -1,0 +1,28 @@
+import type { Config } from 'vike/types'
+import type _ from 'vike-react/config' // Needed for declaration merging of Config
+import type { ApolloClient } from '@apollo/client-react-streaming'
+
+export default {
+  name: 'vike-react-apollo',
+  require: {
+    'vike-react': '>=0.4.18'
+  },
+  Wrapper: 'import:vike-react-apollo/renderer/Wrapper:default',
+  streamIsRequired: true,
+  meta: {
+    ApolloClient: {
+      env: {
+        server: true,
+        client: true
+      }
+    }
+  }
+} satisfies Config
+
+declare global {
+  namespace Vike {
+    interface Config {
+      ApolloClient?: (pageContext: PageContext) => ApolloClient
+    }
+  }
+}