+++ packages/vike-react/src/types/Config.ts
@@ -106,6 +106,7 @@ declare global {
        * https://vike.dev/onAfterRenderClient
        */
       onAfterRenderClient?: (pageContext: PageContextClient) => void
+      Loading?: Loading | ImportString
     }
     interface ConfigResolved {
       Wrapper?: Wrapper[]
@@ -119,6 +120,7 @@ type PlainOrGetter<T> = T | ((pageContext: PageContext) => T)
 export type Head = React.ReactNode | (() => React.ReactNode)
 type Wrapper = (props: { children: React.ReactNode }) => React.ReactNode
 type Layout = Wrapper
+type Loading = { component?: () => React.ReactNode; layout?: () => React.ReactNode }
 
 export type ConfigFromHook = {
   /**